Implemented collisions for Lava actor
The origianl static mesh will act as the damage collision, therefore it
will be able to overlap with other actors.

The copy of the static mesh will act as the floor of the lava which can
be stepped on, therefore it will have to block other actors.

This way, it will give a slight sink in effect, and it could be applied
to different static meshes easily.

TODO: Implement damage functionality of Lava.