feat: volumes saved through game instance
The 'Continue Game' button is no longer enabled when saving volumes in the 'Settings' menu (before actually starting the game as a new player), that data is now stored in 'BFH_Game_Instance'. If the player quits the game, then the volumes are set back to a default value of 0.5, but sound volumes persist between main menu and game level.