Revisions pt. 2
In BP_FireplaceHealingVolume, delete the Tilemap Component (it is not ready for production, so it turns out we shouldn't be using it. Instead, let's put back the fireplace tilemaps everywhere on MAP_Tutorial_Room directly. I'd suggest reverting Tutorial Room to the version on main using https://www.freecodecamp.org/news/git-checkout-file-from-another-branch/ and then manually adding the fireplace healing volumes back wherever the fireplaces exist
In BP_HealingVolume, check if actor implements BP_DamageableEffects in OnBeginOverlap and on OnExitOverlap. If it does, OnBeginOverlap, call SpawnHealingParticles on that actor with HealingParticlesToSpawn and on OnExitOverlap call DespawnHealingParticles on actor
In BP_Violin, implement SpawnHealingParticles to spawn the specific healing particle FX provided and cache the Niagara Particle System Object component created (save it as a variable) to reference when despawning.
In BP_Violin, implement DespawnHealingParticles to despawn the cached healing particle reference or do nothing if cached instance is not valid (check IsValid).
Delete the SpawnHealingParticles function you'd created earlier in BP_Violin
Change FXS_FireplaceHeal to have a pretty long (10+ seconds) default cooldown so that the particles don't disappear until DespawnParticles is explicitly called.

Bugs/todo:
- can't get healing particles to attach to player
- healing particles don't disappear after player reaches max health
- healing doesn't have sfx